+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 - Show who's writing, and who's the recipient
 - Less help text when writing, more concise
 - The blacklist works better (recorded and compared downcased)
-- Added the Convert database command
-- Export blacklist, alias and settings as JSON
+- Added the Convert blacklist command
+- Export blacklist as CSV, aliases and settings as JSON
 
 # 1.0.13